Refactor to have main & favs activities extend same parent
